djangophish/requirements.txt
asgiref==3.2.7
certifi==2020.6.20
Django==3.0.7
django-cleanup==5.0.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
olefile==0.46
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
psycopg2-binary>=2.8